requirements.txt
langchain-groq
langchain
python-dotenv
langchain-community
flask
flask-cors
langchain-core
chromadb
sentence-transformers
langchain-huggingface
pymongo
